featuresディレクトリの内容をappディレクトリへ移動する
