+++ src/app/_hooks/usePosts.ts
@@ -1,8 +1,8 @@
 "use client";
 
 import { useState, useEffect } from "react";
-import { API_BASE_URL } from "@/features/blog/lib/constants";
-import { Post } from "@/features/blog/types/types";
+import { API_BASE_URL } from "@/app/_lib/constants";
+import { Post } from "@/app/_types/types";
 
 function usePosts() {
   const [posts, setPosts] = useState<Post[]>([]);
