+++ src/app/_types/index.ts
@@ -4,3 +4,12 @@ export interface FieldBaseProps {
   label: string;
   error?: FieldError;
 }
+
+export type Post = {
+  id: number;
+  title: string;
+  thumbnailUrl: string;
+  createdAt: string;
+  categories: string[];
+  content: string;
+};
