--- src/features/blog/types/types.ts
@@ -1,8 +1,0 @@
-export type Post = {
-  id: number;
-  title: string;
-  thumbnailUrl: string;
-  createdAt: string;
-  categories: string[];
-  content: string;
-};
